# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  model.InvoiceRequest:
    properties:
      amount:
        type: integer
      cust_email:
        type: string
      cust_name:
        type: string
      dueInDays:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.ErrorBody'
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for Project VioPal
  title: VioPal
  version: "1.0"
paths:
...
  /api/v1/admin/stats:
    get:
      consumes:
      - application/json
      description: Retrieve invoice statistics filtered by merchant_id and created
        date range
      parameters:
      - description: Merchant ID
        in: query
        name: merchant_id
        type: integer
      - description: Start created date (YYYY-MM-DD)
        in: query
        name: start_created_date
        type: string
      - description: End created date (YYYY-MM-DD)
        in: query
        name: end_created_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invoice statistics
      tags:
      - Admin
...
  /api/v1/invoices:
    get:
      consumes:
      - application/json
      description: Get paginated invoice list with optional filters (status, customer
        name, email)
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: perpage
        type: integer
      - description: Invoice status (PENDING, PAID, EXPIRED)
        in: query
        name: status
        type: string
      - description: Customer name (partial match)
        in: query
        name: cust_name
        type: string
      - description: Customer email (partial match)
        in: query
        name: cust_email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List invoices
      tags:
      - Invoice
    post:
      consumes:
      - application/json
      description: Create new invoice request for merchant
      parameters:
      - description: Invoice Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create invoice
      tags:
      - Invoice
  /api/v1/invoices/{code}:
    get:
      description: Get invoice detail using invoice code for authenticated user
      parameters:
      - description: Invoice code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invoice by code
      tags:
      - Invoice
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"